requirements.txt
charset-normalizer
numpy
openpyxl
pandas
pydantic
pydantic_core
python-dateutil
python-dotenv
python-telegram-bot
tenacity
watchdog